+++ gcw-backend/gcw-services/controllers/serviceController.js
@@ -1,39 +1,59 @@
 const Service = require('../models/Service');
 
 const getServices = (req, res) => {
-    Service.find({}).then((services) => {
-        res.send(services);
-    }).catch((err) => {
-        res.json({ message: err.message });
-    });
+    if (req.userType && req.userType === "Admin") {
+        Service.find({}).then((services) => {
+            res.send(services);
+        }).catch((err) => {
+            res.json({ error: err.message });
+        });
+    } else {
+        res.json({ error: "Invalid User" });
+    }
 };
 
 const addService = (req, res) => {
-    Service.create(req.body).then((service) => {
-        res.send(service);
-    }).catch((err) => {
-        res.json({ message: err.message });
-    });
+    if (req.userType && req.userType === "Admin") {
+        let newService = req.body;
+        newService.createdBy = req.userId;
+        Service.create(newService).then((service) => {
+            res.send(service);
+        }).catch((err) => {
+            res.json({ error: err.message });
+        });
+    } else {
+        res.json({ error: "Invalid User" });
+    }
 };
 
 const updateService = (req, res) => {
-    Service.findByIdAndUpdate({ _id: req.params.id }, req.body).then(() => {
-        Service.findById({ _id: req.params.id }).then((service) => {
-            res.send(service);
+    if (req.userType && req.userType === "Admin") {
+        let updatedService = req.body;
+        updatedService.updatedAt = Date.now();
+        Service.findByIdAndUpdate({ _id: req.params.id }, updatedService).then(() => {
+            Service.findById({ _id: req.params.id }).then((service) => {
+                res.send(service);
+            }).catch((err) => {
+                res.json({ error: err.message });
+            });
         }).catch((err) => {
-            res.json({ message: err.message });
+            res.json({ erorr: err.message });
         });
-    }).catch((err) => {
-        res.json({ message: err.message });
-    });
+    } else {
+        res.json({ error: "Invalid User" });
+    }
 };
 
 const deleteService = (req, res) => {
-    Service.findByIdAndDelete({ _id: req.params.id }).then(() => {
-        res.json({ message: "Service Deleted successfully" });
-    }).catch((err) => {
-        res.json({ message: err.message });
-    });
+    if (req.userType && req.userType === "Admin") {
+        Service.findByIdAndDelete({ _id: req.params.id }).then(() => {
+            res.json({ message: "Service Deleted successfully" });
+        }).catch((err) => {
+            res.json({ error: err.message });
+        });
+    } else {
+        res.json({ error: "Invalid User" });
+    }
 };
 
 module.exports = { getServices, addService, updateService, deleteService };
